+++ handlers/admin/vars.go
@@ -4,7 +4,7 @@ import (
 	"database/sql"
 
 	"github.com/arran4/goa4web/core"
-	"github.com/arran4/goa4web/pkg/server"
+	"github.com/arran4/goa4web/internal/server"
 )
 
 // ConfigFile points to the application's config file.
