+++ internal/app/run.go
@@ -25,8 +25,8 @@ import (
 	csrfmw "github.com/arran4/goa4web/internal/middleware/csrf"
 	notifications "github.com/arran4/goa4web/internal/notifications"
 	routerpkg "github.com/arran4/goa4web/internal/router"
+	"github.com/arran4/goa4web/internal/server"
 	startup "github.com/arran4/goa4web/internal/startup"
-	"github.com/arran4/goa4web/pkg/server"
 	"github.com/arran4/goa4web/runtimeconfig"
 	"github.com/gorilla/mux"
 	"github.com/gorilla/sessions"
